rename sample package
