+++ sample/src/androidTest/java/com/aiqcomponents/sample/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.aiqfome.aiqandroid;
+package com.aiqcomponents.sample;
 
 import android.content.Context;
 
@@ -22,6 +22,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
 
-        assertEquals("com.aiqfome.aiqandroid", appContext.getPackageName());
+        assertEquals("com.aiqcomponents.sample", appContext.getPackageName());
     }
 }
