+++ sample/src/main/java/com/aiqcomponents/sample/Country.java
@@ -1,4 +1,4 @@
-package com.aiqfome.aiqandroid;
+package com.aiqcomponents.sample;
 
 import android.graphics.drawable.Drawable;
