+++ sample/src/test/java/com/aiqcomponents/sample/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.aiqfome.aiqandroid;
+package com.aiqcomponents.sample;
 
 import org.junit.Test;
